requirements.txt
fastapi==0.95.2           # FastAPI framework
uvicorn[standard]==0.22.0 # ASGI server to run FastAPI
sqlalchemy==1.4.48        # ORM for PostgreSQL
psycopg2-binary==2.9.8    # PostgreSQL database driver
pydantic==1.10.9          # Data validation and settings management
python-dotenv==1.0.0      # Load environment variables from a .env file
openai==0.27.8            # OpenAI API integration
alembic==1.11.2
requests==2.31.0
